Check if identifier exists

This function quickly checks to see if an identifier exists. It does a quick head request to skip the overhead of item metadta retrieval. This will also return FALSE if the identifier exists but is associated with an item that is unavailable due to permission restrictions.

Usage

identifier_exists(sb_id, ...)

Arguments

sb_id

An sbitem object or a character ScienceBase ID corresponding to the item

...

Additional parameters are passed on to GET, POST, HEAD, PUT.

Value

Logical, TRUE or FALSE

Examples

# \donttest{
# identifier exists
identifier_exists(sb_id = "57976a0ce4b021cadec97890")
#> [1] TRUE

# identifier does not exist
identifier_exists(sb_id = "aaaaaaakkkkkkkbbbbbb")
#> Request failed [404]. Retrying in 1 seconds...
#> Request failed [404]. Retrying in 3.3 seconds...
#> [1] FALSE
# }