Get an item's parent ID

Retrieves the parent of a supplied item based on the ScienceBase item tree hierarchy.

Usage

item_get_parent(sb_id, ...)

Arguments

sb_id

An sbitem object or a character ScienceBase ID corresponding to the item

...

Additional parameters are passed on to GET, POST, HEAD, PUT.

Value

An item object representing the parent of the supplied item.

Examples

# \donttest{
item_get_parent("57976a0ce4b021cadec97890")
#> <ScienceBase Item> 
#>   Title: Select Topgraphic Attributes for NHDPlus Version 2.1 Reach Catchments and Modified Network Routed Upstream Watersheds for the Conterminous United States
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):  2016-07-15T19:59:10Z / 2023-08-02T16:01:24Z
#>   Children: TRUE
#>   Item ID: 5789408ee4b0c1aacab7770b
#>   Parent ID: 5669a79ee4b08895842a1d47

item_get_parent(item_get("57976a0ce4b021cadec97890"))
#> <ScienceBase Item> 
#>   Title: Select Topgraphic Attributes for NHDPlus Version 2.1 Reach Catchments and Modified Network Routed Upstream Watersheds for the Conterminous United States
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):  2016-07-15T19:59:10Z / 2023-08-02T16:01:24Z
#>   Children: TRUE
#>   Item ID: 5789408ee4b0c1aacab7770b
#>   Parent ID: 5669a79ee4b08895842a1d47
# }