Update many SB items with new metadata

A method to update multiple ScienceBase items with a single call and a single HTTP service request. Can be useful for improving performance of updating a large number of items at once.

Usage

items_update(sb_id, info, ...)

Arguments

sb_id

An sbitem object or a character ScienceBase ID corresponding to the item

info

list of metadata info (key-value pairs) to change on the item

...

Additional parameters are passed on to PUT

Value

One or more objects of class sbitem in a list

Details

If length of sb_id > 1, then length of info input must be the same

Examples

if (FALSE) {
# helper function to make a random name
aname <- function() paste0(sample(letters, size = 5, replace = TRUE), collapse = "")

res <- items_create(user_id(), title = c(aname(), aname())) 
out <- items_update(res, info = list( list(title = aname()), list(title = aname()) ) )
vapply(out, "[[", "", "title")
}