Query SB for specific data type

Queries ScienceBase for items with matching datatype.

Usage

query_sb_datatype(datatype, ..., limit = 20)

Arguments

datatype

Character string indicating datatype. See sb_datatypes for full list of available datatypes.

...

Additional parameters are passed on to GET

limit

Maximum number of returned items. Will do paging to retrieve results when limit is over 1000. Use with caution, queries 10k results are slow.

Value

A list of sbitem objects. List of length 0 means no matches were found.

Examples

# \donttest{
#query for items with WFS Layer data
query_sb_datatype('Static Map Image')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: GLO & BLM Survey Plats for Wyoming
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4813e4b07f02db4da961
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: Downloadable Maps for Presentations
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4884e4b07f02db51840f
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: The Washington-British Columbia Transboundary Climate-Connectivity Project
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 57b3b8e7e4b03bcb0103980a
#>   Parent ID: 5318c6bae4b0ae6e9d5a3bb1
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Ground Penetrating Radar Soil Suitability Map for Wyoming
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e47d9e4b07f02db4b53af
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Wyoming State Archives
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4a6ee4b07f02db63fc40
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Library of Congress Geography and Maps Division Map Collections
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4a70e4b07f02db641399
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Wyoming State Historical Society
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e479ae4b07f02db48ffaa
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Gas Certificated Areas
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4b11e4b07f02db6a2412
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Electric Utility Map
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4acfe4b07f02db6801c4
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Telephone Certificated Areas
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e47a5e4b07f02db497e2c
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 

#query for US Topo maps
query_sb_datatype('Map Service')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Uzbekistan, Gazli: 1976-2015
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a95c1f5e4b06990606a7f97
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: XYZ files of LiDAR and sonar data collected at Lake Superior at Minnesota Point, Duluth, MN, August 2019
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5e3ad833e4b0edb47bddabdb
#>   Parent ID: 5dbc9a51e4b06957974edf4c
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: Chesapeake Bay Watershed 2011 Edition Land Cover Data Release
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 557b238fe4b0c350d7b9abb8
#>   Parent ID: 5575e4b7e4b05803ad1ec91d
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Chile, Valparaiso: 2001-2017
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5aa937ebe4b0b1c392f12b6c
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: China, Jiashi: 1977-2004
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a8c4d09e4b00f54eb440461
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Federal Fluid Minerals Leases (Oil and Gas) for the Bureau of Land Management
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 576430c0e4b07657d19ba852
#>   Parent ID: 57642d9ce4b07657d19ba840
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Undiscovered Oil and Gas Resources for State of Montana
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 579b86cde4b0589fa1c981be
#>   Parent ID: 57642d9ce4b07657d19ba840
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Proportion July - Sep) - 2020-2050 - RCP8.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59cea7cbe4b05fe04cc05ef5
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Temperature (Mean: Annual) - 2020-2050 - RCP8.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59cea9f1e4b05fe04cc05f16
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Proportion May - Oct) - 1980-2010
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59ceaa13e4b05fe04cc05f18
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Proportion May - Oct) - 2070-2100 - RCP4.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59ceaa6de4b05fe04cc05f20
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Proportion May - Oct) - 2020-2050 - RCP4.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59ceaaf7e4b05fe04cc05f2e
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Mean: Apr - June) - 2070-2100 - RCP4.5 - Max
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59ceac55e4b05fe04cc05f49
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Mean: Dec - Mar) - 2020-2050 - RCP4.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59ceb113e4b05fe04cc061d7
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Saudi Arabia, Aqaba: 1993-2016
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5ac58824e4b0e2c2dd0c9dca
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: USA, Utah, Magna: 1978-2020
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5e790159e4b01d509270e52b
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Conservation Planning for the Colorado River in Utah - Access to the Site for Relative Cost of Restoration Model
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 587d020ee4b07af201111b93
#>   Parent ID: 582f4d1be4b04d580bd5662d
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: Connecticut Unpublished Surficial Geology Maps
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5228d3e2e4b06291bed806d2
#>   Parent ID: 5220ebd2e4b0645fc25fbb8b
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Temperature (Mean: Annual) - 2020-2050 - RCP4.5 - Max
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59cea90ee4b05fe04cc05f0a
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Conservation Planning for the Colorado River in Utah - Bottomland Boundary of the Colorado River Divided at Homogeneous River Reaches
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5aa6d3fbe4b0b1c392ed6791
#>   Parent ID: 52716d93e4b0f7a10664fe3e
#> 
# }