R interface to ScienceBase

This package provides a rich interface to USGS's ScienceBase https://www.sciencebase.gov/ - a data cataloging and collaborative data management platform. For further information, see the sbtools manuscript here.

Functions are inluded for searching for data, retrieving, creating, and updating datasets.

Details

Functionality in this package allows all users to query ScienceBase for data using a variety of metadata types (query_sb_text, query_sb_doi, query_sb_spatial). Items and associated information can be requested by item_get including item parents item_get_parent and children item_list_children. Data and attached files can be accessed for all available items through provided functionality (e.g., item_file_download).

Authentication

See the function authenticate_sb to authenticate. You'll be required to pass in your ScienceBase username and password.

Authenticated users can create, update, and remove items (item_list_children, item_list_children, item_create, item_update, item_rm).

Feedback

Report any feedback or bugs at https://github.com/DOI-USGS/sbtools/issues