Checks current session and re-authenticates if necessary

Checks the state of your Sciencebase session, re-authenticates if the session is expired, and simply renews if the session is active.

Usage

session_renew(password, ..., username)

Arguments

password

The password to use, if needed, to renew the session.

...

Any additional parameters are currently ignored.

username

Optional. Used only to confirm that the current username is what you expect; if you want to switch usernames, use authenticate_sb() instead of this function.

Value

Returns the session object.

Examples

if (FALSE) { # \dontrun{
# an empty call is sufficient if the session is current, 
# but will break if haven't been logged in before
session_renew()

# include a password if session may be expired
session_renew('newpass')

# optionally confirm the value of the current username
session_renew(username='olduser@usgs.gov', 'newpass')
} # }