Return IDs for all child items

Returns a list of child IDs for a ScienceBase item

Usage

item_list_children(sb_id, fields = c("id", "title"), ..., limit = 20)

Arguments

sb_id

An sbitem object or a character ScienceBase ID corresponding to the item

fields

A character vector of requested data fields. Defaults to 'id' and 'title'. Full list of possible fields is available online in SB documentation.

...

Additional parameters are passed on to GET, POST, HEAD, PUT.

limit

Max children returned.

Value

List of sbitem for each child item.

Examples

if (FALSE) {
item_list_children(user_id())
}

# \donttest{
item_list_children(as.sbitem('5060b03ae4b00fc20c4f3c8b'))
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Madison CO-Alphabetical index; Document # 421874; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfee4b005831885ec60
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Sweetwater Tract Hole SW-8; Document # 421215; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c16e4b005831885fbc6
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Shelden estate Hole 52W114; Document # 422859; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c1be4b005831885ffb0
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 87
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233ce
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 79
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233be
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 67
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233a8
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 51
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285ee4b002407f22339c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 4235
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285ee4b002407f223390
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 4224
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285de4b002407f22337a
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Rhodes Tract #R15, drill hole record - test geophysical survey; Document # 419846; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c06e4b005831885f178
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Rhodes Tract #R100 & 101, drill hole record - test geophysical survey; Document # 419873; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c06e4b005831885f16e
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; O'Brien Tract Hole 61, drill hole record - test geophysical survey; Document # 419935; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c05e4b005831885f09c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; O'Brien Tract Hole 31, drill hole record - test geophysical survey; Document # 419905; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c04e4b005831885f052
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; John Newcum Tract Hole JN86, drill hole record - test geophysical survey; Document # 420018; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfce4b005831885eb6c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; John Newcum Tract Hole JN-95, drill hole record - test geophysical survey; Document # 420009; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfce4b005831885eb50
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 29, Drill Hole record - test geophysical survey; Document # 419946; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b7ae4b005831885e96d
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 21, Drill Hole record - test geophysical survey; Document # 419954; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b7ae4b005831885e95d
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 8, Drill Hole record - test geophysical survey; Document # 419967; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b79e4b005831885e93f
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Gilbert O Combs 6, Drill Hole Record; Document # 420300; Bollinger County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b69e4b005831885e2e9
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Sexton, Michael J. Tract SX-5; Document # 421227; Bollinger County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b6ae4b005831885e333
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
item_list_children(item_get('5060b03ae4b00fc20c4f3c8b'))
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Madison CO-Alphabetical index; Document # 421874; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfee4b005831885ec60
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Sweetwater Tract Hole SW-8; Document # 421215; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c16e4b005831885fbc6
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Shelden estate Hole 52W114; Document # 422859; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c1be4b005831885ffb0
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 87
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233ce
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 79
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233be
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 67
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285fe4b002407f2233a8
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 51
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285ee4b002407f22339c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 4235
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285ee4b002407f223390
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Missouri Geological Survey Mineral Resource Microfilm Reels, Reel Number 4224
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c1285de4b002407f22337a
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Rhodes Tract #R15, drill hole record - test geophysical survey; Document # 419846; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c06e4b005831885f178
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Rhodes Tract #R100 & 101, drill hole record - test geophysical survey; Document # 419873; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c06e4b005831885f16e
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; O'Brien Tract Hole 61, drill hole record - test geophysical survey; Document # 419935; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c05e4b005831885f09c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; O'Brien Tract Hole 31, drill hole record - test geophysical survey; Document # 419905; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11c04e4b005831885f052
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; John Newcum Tract Hole JN86, drill hole record - test geophysical survey; Document # 420018; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfce4b005831885eb6c
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; John Newcum Tract Hole JN-95, drill hole record - test geophysical survey; Document # 420009; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11bfce4b005831885eb50
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 29, Drill Hole record - test geophysical survey; Document # 419946; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b7ae4b005831885e96d
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 21, Drill Hole record - test geophysical survey; Document # 419954; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b7ae4b005831885e95d
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Glaves Tract Hole GL 8, Drill Hole record - test geophysical survey; Document # 419967; Madison County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b79e4b005831885e93f
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Gilbert O Combs 6, Drill Hole Record; Document # 420300; Bollinger County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b69e4b005831885e2e9
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Missouri Mineral Resource Microfilm File; Sexton, Michael J. Tract SX-5; Document # 421227; Bollinger County
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 50c11b6ae4b005831885e333
#>   Parent ID: 5060b03ae4b00fc20c4f3c8b
#> 
# }