Query SB for specific data type

Queries ScienceBase for items with matching datatype.

Usage

query_sb_datatype(datatype, ..., limit = 20)

Arguments

datatype

Character string indicating datatype. See sb_datatypes for full list of available datatypes.

...

Additional parameters are passed on to GET

limit

Maximum number of returned items. Will do paging to retrieve results when limit is over 1000. Use with caution, queries 10k results are slow.

Value

A list of sbitem objects. List of length 0 means no matches were found.

Examples

# \donttest{
#query for items with WFS Layer data
query_sb_datatype('Static Map Image')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Downloadable Maps for Presentations
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4884e4b07f02db51840f
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: GLO & BLM Survey Plats for Wyoming
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4813e4b07f02db4da961
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: The Washington-British Columbia Transboundary Climate-Connectivity Project
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 57b3b8e7e4b03bcb0103980a
#>   Parent ID: 5318c6bae4b0ae6e9d5a3bb1
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Ground Penetrating Radar Soil Suitability Map for Wyoming
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e47d9e4b07f02db4b53af
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Library of Congress Geography and Maps Division Map Collections
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4a70e4b07f02db641399
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Wyoming State Archives
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4a6ee4b07f02db63fc40
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Wyoming State Historical Society
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e479ae4b07f02db48ffaa
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Gas Certificated Areas
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4b11e4b07f02db6a2412
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Electric Utility Map
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e4acfe4b07f02db6801c4
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Wyoming Public Service Commission Telephone Certificated Areas
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 4f4e47a5e4b07f02db497e2c
#>   Parent ID: 4f4e4772e4b07f02db47e22e
#> 

#query for US Topo maps
query_sb_datatype('Map Service')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: 0.2-second spectral response acceleration (5% of critical damping) with a 1% probability of exceedance in 1 year for the Western United States
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a90ade1e4b0699060675016
#>   Parent ID: 5a8c9b70e4b069906054dfab
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: Uzbekistan, Gazli: 1976-2015
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a95c1f5e4b06990606a7f97
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: XYZ files of LiDAR and sonar data collected at Lake Superior at Minnesota Point, Duluth, MN, August 2019
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5e3ad833e4b0edb47bddabdb
#>   Parent ID: 5dbc9a51e4b06957974edf4c
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Chesapeake Bay Watershed 2011 Edition Land Cover Data Release
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 557b238fe4b0c350d7b9abb8
#>   Parent ID: 5575e4b7e4b05803ad1ec91d
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Chile, Valparaiso: 2001-2017
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5aa937ebe4b0b1c392f12b6c
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Proportion of Low and Black Sagebrush Land Cover (5-km scale) in the Wyoming Basins Ecoregional Assessment area
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 54341d08e4b095098ca8be86
#>   Parent ID: 5fc56875d34e4b9faad88921
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Mixed Shrubland Land Cover in the Wyoming Basins Ecoregional Assessment area
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 54342fb3e4b095098ca8d680
#>   Parent ID: 5fc56875d34e4b9faad88921
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Oil and Gas Well Density in the Western United States
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5436fda5e4b0a4f4b46a832b
#>   Parent ID: 5fd90d69d34e30b9123cbc21
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: CW89. View east toward flood-damaged travertine pools below Havasu Falls
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5c75a755e4b0fe48cb4a5513
#>   Parent ID: 5c6200cde4b0fe48cb32ff6d
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: EI09. View north toward basalt dike in Hakatai Shale
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5c75a872e4b0fe48cb4a570c
#>   Parent ID: 5c6200cde4b0fe48cb32ff6d
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: China, Jiashi: 1977-2004
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a8c4d09e4b00f54eb440461
#>   Parent ID: 59fb91fde4b0531197b16ac7
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Elevation artifacts in digital bathymetric and topographic models for United States east (east_cdem_v1.tif) and west (west_cdem_v.tif) coasts (polygon shapefile, geographic, NAD83)
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 57ebe895e4b09082500ef4a8
#>   Parent ID: 57dc6e86e4b090824ffe16eb
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: Federal Fluid Minerals Leases (Oil and Gas) for the Bureau of Land Management
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 576430c0e4b07657d19ba852
#>   Parent ID: 57642d9ce4b07657d19ba840
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Undiscovered Oil and Gas Resources for State of Montana
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 579b86cde4b0589fa1c981be
#>   Parent ID: 57642d9ce4b07657d19ba840
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Low-altitude aerial imagery obtained with unmanned aerial systems (UAS) flights over Coast Guard Beach, Nauset Spit, Nauset Inlet, and Nauset Marsh, Cape Cod National Seashore, Eastham, Massachusetts on 1 March 2016 (JPEG images)
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 579f8532e4b0589fa1cd54f7
#>   Parent ID: 579a6face4b0589fa1c8718c
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: 3) Comparison with previous models
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59bc43ebe4b091459a59137d
#>   Parent ID: 58795a8ce4b04df303d97ed8
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Shorelines of the Florida north (FLnorth) coastal region used in shoreline change analysis
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 582ca519e4b04d580bd3794a
#>   Parent ID: 582ca495e4b04d580bd378cd
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: Data from mercury intrusion porosimetry conducted on samples of a mudstone underlying the Naval Air Warfare Center, West Trenton, New Jersey
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5898a1a3e4b0efcedb7073e8
#>   Parent ID: 5474ec49e4b04d7459a7eab2
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Bathymetric Survey of the Ichilo and Sajta Rivers, near Puerto Villarroel, Bolivia, May 23-24, 2019
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5d961f59e4b0c4f70d110e39
#>   Parent ID: 5474ec49e4b04d7459a7eab2
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Precipitation (Proportion July - Sep) - 2020-2050 - RCP8.5 - Min
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 59cea7cbe4b05fe04cc05ef5
#>   Parent ID: 5850549ae4b0f24ebfd9368f
#> 
# }