Query SB for items containing specific text

Queries for ScienceBase items that have matching text in the title or description

Usage

query_sb_text(text, ..., limit = 20)

Arguments

text

Text string for search

...

Additional parameters are passed on to GET

limit

Maximum number of returned items. Will do paging to retrieve results when limit is over 1000. Use with caution, queries 10k results are slow.

Value

A list of sbitem objects. List of length 0 means no matches were found.

Examples

# \donttest{
#query for a person's name
query_sb_text('Luna Leopold')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Dr. Luna B. Leopold, Senior Research Hydrologist, former Chief Hydrologist (1957 - 1966); member National Academy of Sciences.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dda2dce4b0f72b4471dff5
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: East Branch Brandywine Creek, Pennsylvania
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5fe3caadd34ea5387deb4b22
#>   Parent ID: 5fe39194d34ea5387deb4933
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: Young Pits, Sisters Area, Maryland
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5fea160bd34ea5387ded6b7c
#>   Parent ID: 5fe39194d34ea5387deb4933
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Powder River at Bulkley's Ranch. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7946e4b0f72b4471ad5d
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Frost wedge and congeliturbate on west end of Arvada Bridge, Powder River. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7944e4b0f72b4471ad57
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Frost wedge and congeliturbate on west end of Arvada Bridge. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7945e4b0f72b4471ad5b
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment of Arvada Bridge. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7948e4b0f72b4471ad6b
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment of Arvada Bridge. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7948e4b0f72b4471ad69
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Tributary to Powder River. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7944e4b0f72b4471ad55
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Channel of Belle Fourche River near Carlile Bridge. Crook County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7942e4b0f72b4471ad49
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: Hasta Luego Draw near Billings, Montana, 18.6 miles southeast of Billings on Highway 87, north of road.  Yellowstone County, Montana. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7949e4b0f72b4471ad6f
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment of Arvada Bridge across Powder River. Sheridan County, Wyoming.1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7947e4b0f72b4471ad67
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: High terrace standing nearly 80 feet above Powder Rive. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7944e4b0f72b4471ad53
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Repeat photograph of Puerco River at Hogback, two miles east of Gallup, McKinley County, New Mexico. 1946.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd793fe4b0f72b4471ad39
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment of Arvada Bridge across Powder River. Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7948e4b0f72b4471ad6d
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: Mexican Springs Wash, one-quarter mile above Chuska Wash; showing caliche band about five feet below surface. San Juan County, New Mexico. 1949.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd794be4b0f72b4471ad7f
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7947e4b0f72b4471ad63
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7946e4b0f72b4471ad5f
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Congeliturbate site on road cut at west abutment. Sheridan County, Wyoming. 1950.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd7946e4b0f72b4471ad61
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Retake of Dutton 21, Pyramid Rock (ratio 1:3.6). McKinley County, New Mexico. 1946.
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 51dd793be4b0f72b4471ad2b
#>   Parent ID: 519ba0a3e4b0e4e151ef5dd9
#> 

#query for one of the old river gaging stations
query_sb_text('Lees Ferry')
#> [[1]]
#> <ScienceBase Item> 
#>   Title: Lees Ferry Area
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751de4b00f54eb1e44e8
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[2]]
#> <ScienceBase Item> 
#>   Title: Rockfall near Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787603e4b00f54eb1e5942
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[3]]
#> <ScienceBase Item> 
#>   Title: Rockfall near Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787603e4b00f54eb1e5940
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[4]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751de4b00f54eb1e44ec
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[5]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e44f0
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[6]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e4502
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[7]]
#> <ScienceBase Item> 
#>   Title: Lees Ferry Area
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751de4b00f54eb1e44ea
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[8]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e44ee
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[9]]
#> <ScienceBase Item> 
#>   Title: Flood mitigation at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787602e4b00f54eb1e5936
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[10]]
#> <ScienceBase Item> 
#>   Title: Relics at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787601e4b00f54eb1e592e
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[11]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787602e4b00f54eb1e5932
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[12]]
#> <ScienceBase Item> 
#>   Title: Hoodoos at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787605e4b00f54eb1e5950
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[13]]
#> <ScienceBase Item> 
#>   Title: Shinarump rockfall at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787607e4b00f54eb1e5966
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[14]]
#> <ScienceBase Item> 
#>   Title: Channel Erosion at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e4504
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[15]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e44fe
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[16]]
#> <ScienceBase Item> 
#>   Title: Colorado River at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a78751ee4b00f54eb1e4500
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[17]]
#> <ScienceBase Item> 
#>   Title: Hoodoos near Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787605e4b00f54eb1e5952
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[18]]
#> <ScienceBase Item> 
#>   Title: 09380000_Colorado River at Lees Ferry, AZ
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5f1603f182cef313ed819d36
#>   Parent ID: 5f15fdc982cef313ed8199c9
#> 
#> [[19]]
#> <ScienceBase Item> 
#>   Title: Flood/erosion mitigation at Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787602e4b00f54eb1e5934
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
#> [[20]]
#> <ScienceBase Item> 
#>   Title: Vermilion Cliffs near Lees Ferry
#>   Creator/LastUpdatedBy:      / 
#>   Provenance (Created / Updated):   / 
#>   Children: 
#>   Item ID: 5a787605e4b00f54eb1e5948
#>   Parent ID: 4f4e49d8e4b07f02db5df316
#> 
# }